refactor: separate out image properties
