+++ frontend/src/components/TextProperties.vue
